8/7/2024 - 11:53 am
Added additional score functionality, calculating for all client defined metrics. 
To Do:
Need to change double values for scores so each category is weighted the same when accounting for the percentage values for each. 
Need to format the output so only the top 10% of each category is written to the file.